+++ R/utils.R
@@ -49,17 +49,64 @@ get_zoltar_project_url <- function(hub = c("US", "ECDC"),
   return(project_url)
 }
 
+
+#' Get FIPS or CBSA codes from county or metropolitan area locations
+#'
+#' @param data a vector of strings for fips code, CBSA codes, location names
+#' such as "Hampshire County, MA", "Alabama", "United Kingdom".
+#' A US county location names must include state abbreviation.
+#' @param hub character vector, where the first element indicates the hub
+#' from which to load forecasts. Possible options are `"US"`, `"ECDC"` and `"FluSight"`
+#' @return A vector of FIPS or CBSA codes
+#' @export
+name_to_fips <- function(data, hub = c("US", "ECDC")){
+  if (is.null(data)){
+    return (NULL)
+  }
+  
+  if (hub[1] == "US") {
+    fips_tmp <-  covidHubUtils::hub_locations %>%
+      dplyr::mutate(full_location_name = fips)
+    locations <- dplyr::bind_rows(covidHubUtils::hub_locations, fips_tmp)
+    if(!all(data %in% locations$full_location_name)){
+      stop("Error in name_to_fips: Please provide valid location name.eg: Bullock County,AL")
+    }
+    return (locations[locations$full_location_name %in% data, ]$fips)
+  } else if (hub[1] == "ECDC") {
+    location_tmp <- covidHubUtils::hub_locations_ecdc %>%
+      dplyr::mutate(location_name = location)
+    locations <- dplyr::bind_rows(covidHubUtils::hub_locations_ecdc, location_tmp)
+    if(!all(data %in% locations$location_name)){
+      stop("Error in name_to_fips: Please provide valid location name.")
+    }
+    return(locations[locations$location_name %in% data, ]$location)
+  } else if (hub[1] == "FluSight") {
+    fips_tmp <-  covidHubUtils::hub_locations_flusight %>%
+      dplyr::mutate(location_name = fips)
+    fips_tmp[fips_tmp$location_name == "US",]$location_name <- "United States"
+    locations <- dplyr::bind_rows(covidHubUtils::hub_locations_flusight, fips_tmp)
+    if(!all(data %in% locations$location_name)){
+      stop("Error in name_to_fips: Please provide valid location name.eg: Bullock County,AL")
+    }
+    return (locations[locations$location_name %in% data, ]$fips)
+  }
+} 
+
+
+
+
 #' Add stored information about location name and population by joining the
 #' data with the correct stored location data
 #'
 #' @param data data frame to append location data
 #' @param hub character vector, where the first element indicates the hub
-#' from which to load forecasts. Possible options are `"US"` and `"ECDC"`
+#' from which to load forecasts.
+#' Possible options are `"US"`, `"ECDC"` and `"FluSight"`.
 #'
 #' @return original data with corresponding location information
 #' @export
 join_with_hub_locations <- function(data,
-                                    hub = c("US", "ECDC")) {
+                                    hub = c("US", "ECDC", "FluSight")) {
   if (hub[1] == "US") {
     data <- dplyr::left_join(data,
       covidHubUtils::hub_locations,
@@ -69,6 +116,11 @@ join_with_hub_locations <- function(data,
     data <- dplyr::left_join(data,
       covidHubUtils::hub_locations_ecdc,
       by = c("location")
+    )
+  } else if (hub[1] == "FluSight") {
+    data <- dplyr::left_join(data,
+                             covidHubUtils::hub_locations_flusight,
+                             by = c("location" = "fips")
     )
   }
   return(data)